curat-ai-demo/pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Ines Chami",email = "ines.chami22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "logfire[httpx] (>=3.23.0,<4.0.0)",
    "mcp-database-server (>=0.1.4,<0.2.0)",
    "pydantic-ai (>=0.4.0,<0.5.0)",
    "pydantic-ai-slim[anthropic] (>=0.4.0,<0.5.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "rich (>=14.0.0,<15.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"
flake8 = "^7.3.0"

[tool.poetry]
package-mode = false